refactor: quiet down main
