+++ main.go
@@ -4,10 +4,8 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
-	"strconv"
 
 	"github.com/aws/ec2-macos-init/internal/paths"
 	"github.com/aws/ec2-macos-init/lib/ec2macosinit"
@@ -15,55 +13,7 @@ import (
 
 const (
 	loggingTag         = "ec2-macOS-init"
-	gitHubLink         = "https://github.com/aws/ec2-macos-init"
 )
-
-// Build time variables
-var CommitDate string
-var Version string
-
-// printUsage prints the help text when invalid arguments are provided.
-func printUsage(baseDir string) {
-	fmt.Println("Usage: ec2-macos-init <command> <arguments>")
-	fmt.Println("Commands are:")
-	fmt.Println("    run - Run init using configuration located in " + filepath.Join(baseDir, paths.InitTOML))
-	fmt.Println("    clean - Remove instance history from disk")
-	fmt.Println("    version - Print version information")
-	fmt.Println("For more help: ec2-macos-init <command> -h")
-}
-
-// printVersion prints the output for the version command.
-func printVersion() {
-	fmt.Printf("\nEC2 macOS Init\n"+
-		"Version: %s [%s]\n"+
-		"%s\n"+
-		"Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.\n\n",
-		Version, CommitDate, gitHubLink,
-	)
-}
-
-// checkRootPermissions checks to see if the init application is being run as root.
-func checkRootPermissions() (root bool, err error) {
-	cmd := exec.Command("id", "-u")
-	output, err := cmd.Output()
-	if err != nil {
-		return false, err
-	}
-
-	// Convert for comparison
-	i, err := strconv.Atoi(string(output[:len(output)-1]))
-	if err != nil {
-		return false, err
-	}
-
-	// 0 = root
-	// 501 = non-root user
-	if i == 0 {
-		return true, nil
-	}
-
-	return false, nil
-}
 
 func main() {
 	const baseDir = paths.DefaultBaseDirectory
@@ -80,11 +30,7 @@ func main() {
 	}
 
 	// Check that this is being run by a user with root permissions
-	root, err := checkRootPermissions()
-	if err != nil {
-		logger.Fatalf(71, "Error while checking root permissions: %s", err)
-	}
-	if !root {
+	if !runningAsRoot() {
 		logger.Fatal(64, "Must be run with root permissions!")
 	}
 
@@ -116,4 +62,22 @@ func main() {
 		printUsage(baseDir)
 		os.Exit(2)
 	}
+}
+
+// printUsage prints the help text for this program.
+func printUsage(baseDir string) {
+	fmt.Println("Usage: ec2-macos-init <command> <arguments>")
+	fmt.Println("Commands are:")
+	fmt.Println("    run - Run init using configuration located in " + filepath.Join(baseDir, paths.InitTOML))
+	fmt.Println("    clean - Remove instance history from disk")
+	fmt.Println("    version - Print version information")
+	fmt.Println("For more help: ec2-macos-init <command> -h")
+}
+
+
+// runningAsRoot checks to see if the init application is being run as
+// root.
+func runningAsRoot() bool {
+	// must effectively be root
+	return os.Geteuid() == 0
 }
